% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpress.R
\name{power.est}
\alias{power.est}
\title{power.calc}
\usage{
power.est(in.phyloseq, metadata.var, metadata.vals, start.n = 5,
  alpha = 0.05, beta = 0.95, test.type = "permanova",
  dist.metric = "unifrac", deseq.val = 0, n.rep = 100, burn.in = 10,
  verbose = F, binom = F, switch.val = 0, seed.val = 0)
}
\arguments{
\item{in.phyloseq}{Required. The collection of microbiome samples to use. Required to be a class phyloseq data}

\item{metadata.var}{Required. The column name in the metadata values that is being compared for the power.}

\item{metadata.vals}{Required. Vector with the values to be compared.}

\item{alpha}{Alpha value used to test the microbiome differences. Default is 0.05}

\item{beta}{Beta value used to test proportion of estimated samples that contain significant differences. Default is 0.95}

\item{test.type}{String giving the statistic to test the between the microbiome samples. Default is "permanova"}

\item{dist.metric}{String giving the distance metric to use between the microbiome samples. Default"unifrac"}

\item{deseq.val}{Value to use for the DESeq2 run to trim the taxa. 0 = use all taxa (Default), 0 < x < 1, use the value as a FDR cutoff; x >=1, use the top scoring X values by FDR}

\item{n.rep}{Maximum number of replicates to run before final check if above the beta threshold. Default = 100.}

\item{burn.in}{Value to burning in for the binomial test. Default is 10}

\item{verbose}{Flag to prints intermediate values to the screen. Default = F}

\item{binom}{Flag to use the binomial test to stop low mapping values. Default = F}

\item{switch.val}{sample number at which to automatically swap between sampling and simulation. Default = 0, or only swap once sample number exceeds the number of samples}

\item{seed.val}{Value to set seed value for troubleshooting. Default = 0 (not set)}

\item{start.}{The sample size at which to start the tests. Default is 5.}
}
\value{
Returns a class MPrESS variable with the sample number in addition to runtime information for use in verification
}
\description{
power.calc() calculates the number of samples to dectect a difference between microbiomes with two different metadata values.
}
\examples{
#The following data is from Chavda et al 2016 which phylotyped Enterobacter genomes

#Additional printing and plotting options are availible with plot() and print(). For more information refer to ?plot.ggrasp and ?print.ggrasp
#The following data is from Chavda et al 2016 using Microbiome data from multiple sites in China
# Our example uses the data underpinning the tree shown in Figure 2

library(mpress);
#Loading in the Chinese metadata file
data(ChinaData);
data(SpainData);

#Use summary() to examine the data loaded
summary(china.power)

#Other microbiomes: china.trim from data(ChinaData): unnamed OTUs removed
#                  spain.ibs.trim from

#Use plot() to see the plot of the power and p-value in the different sample numbers
plot(china.power)
}
