% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpress.R
\name{print.mpress}
\alias{print.mpress}
\title{print.mpress
makes a table with the runtime information for the samples investigated an mpress object}
\usage{
\method{print}{mpress}(x)
}
\description{
print.mpress
makes a table with the runtime information for the samples investigated an mpress object
}
\examples{
library(mpress);
#Loading in the microbiome files
data(ChinaData);
data(SpainData);

#Use summary() to examine the data loaded

#Use print() to see the table of the run info in the different sample numbers
print(spain.ibs.power)
}
